Extract image generation code into new image processor for modularity
